requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
requests==2.31.0
gql[requests]==3.4.1
aiohttp==3.9.5
